Commit before bed
