+++ warm-wave/src/app/globals.css
@@ -1,0 +1,10 @@
+.text-opacity-overlay {
+    background: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.5));
+    border-radius: 45%;
+    /*
+    border-radius: 15px;
+    padding: 20px;
+    */
+    padding: 30px 40px;
+    backdrop-filter: blur(5px)
+}
