+++ warm-wave/src/app/layout.js
@@ -24,7 +24,7 @@ export default function RootLayout({ children }) {
       <head>
         <link rel="icon" href="/favicon.ico" />
       </head>
-      <body>
+      <body className="main">
         <ThemeRegistry>
           <Box
             sx={{
