+++ warm-wave/src/components/Homepage/about.js
@@ -1,30 +1,11 @@
 "use client";
 
 import Spacer from "../spacer";
-import { Box, Paper, Typography, Stack, IconButton } from "@mui/material";
+import { Box, Paper, Typography } from "@mui/material";
 import Link from "@mui/material/Link";
 import photoUrls from "@/utils/photo-urls";
-import { ArrowBack, ArrowForward } from "@mui/icons-material";
-import { useEffect, useState } from "react";
 
 export default function About() {
-  const [currentPhotoIndex, setCurrentPhotoIndex] = useState(0);
-
-  function nextPhoto(forward) {
-    if (forward == true) {
-      if (currentPhotoIndex == photoUrls.aboutDesktop.length - 1) {
-        setCurrentPhotoIndex(0);
-      } else {
-        setCurrentPhotoIndex(currentPhotoIndex + 1);
-      }
-    } else if (forward == false) {
-      if (currentPhotoIndex == 0) {
-        setCurrentPhotoIndex(photoUrls.aboutDesktop.length - 1);
-      } else {
-        setCurrentPhotoIndex(currentPhotoIndex - 1);
-      }
-    }
-  }
 
   return (
     <div id="about">
@@ -37,7 +18,7 @@ export default function About() {
       <Box
         sx={{
           width: "100%",
-          backgroundImage: "url(" + photoUrls.aboutDesktop[currentPhotoIndex] + ")",
+          backgroundImage: "url(" + photoUrls.aboutDesktop[0] + ")",
           backgroundRepeat: "no-repeat",
           backgroundPosition: "center",
           backgroundSize: "cover",
@@ -56,7 +37,7 @@ export default function About() {
             sx={{
               width: "60%",
               background:
-                "linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.5))",
+                "linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.5))",
               borderRadius: "15px",
               padding: "2em",
             }}
@@ -65,7 +46,7 @@ export default function About() {
               variant="p"
               component="p"
               align="center"
-              sx={{ fontSize: 18 }}
+              sx={{ fontSize: 18, fontWeight: 500 }}
             >
               Welcome to Warm Wave Studio, founded by Andrew Tyler in 2016. We
               specialize in hi-fi recording, mixing, and mastering, catering to
@@ -98,20 +79,6 @@ export default function About() {
         </Box>
 
         <Spacer height={250} />
-      </Box>
-
-      <Box sx={{ width: "100%", marginTop: "2em" }}>
-        <Stack direction={"row"} justifyContent={"center"}>
-          <IconButton onClick={() => nextPhoto(false)}>
-            <ArrowBack />
-          </IconButton>
-          <Typography component={"b"} variant={"b"}>
-            {currentPhotoIndex}
-          </Typography>
-          <IconButton onClick={() => nextPhoto(true)}>
-            <ArrowForward />
-          </IconButton>
-        </Stack>
       </Box>
     </div>
   );
