+++ warm-wave/src/components/Homepage/header.js
@@ -4,33 +4,15 @@ import { Box, Button, IconButton, Stack, Typography } from "@mui/material";
 import Spacer from "../spacer";
 import photoUrls from "@/utils/photo-urls";
 import { ArrowBack, ArrowForward } from "@mui/icons-material";
-import { useEffect, useState } from "react";
 
 export default function Header() {
-  const [currentPhotoIndex, setCurrentPhotoIndex] = useState(0);
-
-  function nextPhoto(forward) {
-    if (forward == true) {
-      if (currentPhotoIndex == photoUrls.headerDesktop.length - 1) {
-        setCurrentPhotoIndex(0);
-      } else {
-        setCurrentPhotoIndex(currentPhotoIndex + 1);
-      }
-    } else if (forward == false) {
-      if (currentPhotoIndex == 0) {
-        setCurrentPhotoIndex(photoUrls.headerDesktop.length - 1);
-      } else {
-        setCurrentPhotoIndex(currentPhotoIndex - 1);
-      }
-    }
-  }
 
   return (
     <>
       <Box
         sx={{
           width: "100%",
-          backgroundImage: "url(" + photoUrls.headerDesktop[currentPhotoIndex] + ")",
+          backgroundImage: "url(" + photoUrls.headerDesktop[0] + ")",
           backgroundRepeat: "no-repeat",
           backgroundPosition: "center",
           backgroundSize: "cover",
@@ -48,14 +30,7 @@ export default function Header() {
             align="center"
             color="primary"
           >
-            <span
-              style={{
-                background:
-                  "linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.5))",
-                borderRadius: "15px",
-                padding: "20px",
-              }}
-            >
+            <span className="text-opacity-overlay">
               Warm Wave Studio
             </span>
           </Typography>
@@ -64,7 +39,9 @@ export default function Header() {
         <Spacer height={300} />
 
         <Typography fontWeight={500} variant="h3" component="h3" align="center">
-          TYour All-in-One Studio for Perfect Sound
+        <span className="text-opacity-overlay">
+          Your All-in-One Studio for Perfect Sound
+          </span>
         </Typography>
 
         <Spacer height={100} />
@@ -86,19 +63,6 @@ export default function Header() {
         </Box>
 
         <Spacer height={120} />
-      </Box>
-      <Box sx={{ width: "100%", marginTop: "2em" }}>
-        <Stack direction={"row"} justifyContent={"center"}>
-          <IconButton onClick={() => nextPhoto(false)}>
-            <ArrowBack />
-          </IconButton>
-          <Typography component={"b"} variant={"b"}>
-            {currentPhotoIndex}
-          </Typography>
-          <IconButton onClick={() => nextPhoto(true)}>
-            <ArrowForward />
-          </IconButton>
-        </Stack>
       </Box>
     </>
   );
