+++ warm-wave/src/components/Mobile/Homepage/header.js
@@ -1,36 +1,17 @@
 "use client";
 
-import { Box, Button, Typography, Stack, IconButton } from "@mui/material";
+import { Box, Button, Typography } from "@mui/material";
 import Spacer from "@/components/spacer";
 import photoUrls from "@/utils/photo-urls";
-import { ArrowBack, ArrowForward } from "@mui/icons-material";
-import { useEffect, useState } from "react";
 
 export default function MobileHeader() {
-  const [currentPhotoIndex, setCurrentPhotoIndex] = useState(0);
-
-  function nextPhoto(forward) {
-    if (forward == true) {
-      if (currentPhotoIndex == photoUrls.headerMobile.length - 1) {
-        setCurrentPhotoIndex(0);
-      } else {
-        setCurrentPhotoIndex(currentPhotoIndex + 1);
-      }
-    } else if (forward == false) {
-      if (currentPhotoIndex == 0) {
-        setCurrentPhotoIndex(photoUrls.headerMobile.length - 1);
-      } else {
-        setCurrentPhotoIndex(currentPhotoIndex - 1);
-      }
-    }
-  }
 
   return (
     <>
       <Box
         sx={{
           width: "100%",
-          backgroundImage: "url(" + photoUrls.headerMobile[currentPhotoIndex] + ")",
+          backgroundImage: "url(" + photoUrls.headerMobile[0] + ")",
           backgroundRepeat: "no-repeat",
           backgroundPosition: "center",
           backgroundSize: "cover",
@@ -46,12 +27,7 @@ export default function MobileHeader() {
             align="center"
             fontWeight={500}
             color="textPrimary"
-            style={{
-              background:
-                "linear-gradient(rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.5))",
-              borderRadius: "15px",
-              padding: "10px 0px",
-            }}
+            className="text-opacity-overlay"
           >
             Warm Wave Studio
           </Typography>
@@ -59,7 +35,7 @@ export default function MobileHeader() {
 
         <Spacer height={100} />
 
-        <Typography variant="h3" component="h3" align="center" fontWeight={500}>
+        <Typography variant="h3" component="h3" align="center" fontWeight={500} >
         Your All-in-One Studio for Perfect Sound
         </Typography>
 
@@ -82,20 +58,6 @@ export default function MobileHeader() {
         </Box>
 
         <Spacer height={50} />
-      </Box>
-
-      <Box sx={{ width: "100%", marginTop: "2em" }}>
-        <Stack direction={"row"} justifyContent={"center"}>
-          <IconButton onClick={() => nextPhoto(false)}>
-            <ArrowBack />
-          </IconButton>
-          <Typography component={"b"} variant={"b"}>
-            {currentPhotoIndex}
-          </Typography>
-          <IconButton onClick={() => nextPhoto(true)}>
-            <ArrowForward />
-          </IconButton>
-        </Stack>
       </Box>
     </>
   );
